Sort out the merge
